LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A6.3_T5-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Unexpected keyword 'false'. Expected either 'in' or 'of' in enumeration syntax.
S7.9_A6.3_T5

PASS Expected parsing failure

TEST COMPLETE
